Better selection of TinyMCE editor and code cleanup
The ID used to get an instance of the TinyMCE editor was changed from a
static String to a variable. Note that the script still only selects and
works with the first found TinyMCE editor in the page's DOM. Any others
are disregarded. Code was added to account for IDs with dots (".") in
their names, as dots need to be escaped when being used as CSS
selectors.

General code cleanup was performed, including moving all variable
declarations to the top of the script and adding some commenting
throughout the code.